backend_functions/buckets/list_bucket_files_3/requirements.txt
google-cloud-storage
Flask
functions-framework
google-auth
google-auth-oauthlib
google-auth-httplib2
